combining documentSettings and common
