+++ lib/dateTime.js
@@ -1,5 +1,4 @@
-import {getDocumentLocaleSettings, getLanguage} from './documentSettings.js';
-import {merge} from './common.js';
+import {getDocumentLocaleSettings, getLanguage, merge} from './common.js';
 
 const hour24locales = ['da', 'de', 'es', 'fr', 'nl', 'sv', 'tr', 'zh'];
 const mondayFirstDayLocales = ['da', 'de', 'fr', 'nl', 'sv', 'tr'];
