+++ lib/fileSize.js
@@ -1,6 +1,5 @@
-import {getLanguage} from './documentSettings.js';
 import {formatNumber} from './number.js';
-import {validateFormatValue} from './common.js';
+import {getLanguage, validateFormatValue} from './common.js';
 
 export function getFileSizeDescriptor() {
