+++ lib/index.js
@@ -1,3 +1,9 @@
-export {formatDate, formatTime, parseDate, parseTime} from './dateTime.js';
+export {
+	formatDate,
+	formatDateTime,
+	formatTime,
+	parseDate,
+	parseTime
+} from './dateTime.js';
 export {formatFileSize} from './fileSize.js';
 export {formatNumber, formatPercent, parseNumber} from './number.js';
