+++ lib/number.js
@@ -1,5 +1,9 @@
-import {getDocumentLocaleSettings, getLanguage} from './documentSettings.js';
-import {merge, validateInteger, validateFormatValue} from './common.js';
+import {
+	getDocumentLocaleSettings,
+	getLanguage,
+	merge,
+	validateFormatValue
+} from './common.js';
 
 function formatPositiveInteger(value, descriptor) {
 	value = Math.floor(value);
@@ -78,6 +82,22 @@ function validateFormatOptions(options) {
 	}
 
 	return options;
+
+}
+
+function validateInteger(name, value, defaultValue, min, max) {
+
+	if (value === undefined || value === null) {
+		value = defaultValue;
+	}
+	if (typeof value === 'string') {
+		value = parseInt(value);
+	}
+	if (isNaN(value) || typeof value !== 'number' || (min !== undefined && value < min) || (max !== undefined && value > max)) {
+		throw new RangeError(name + ' value is out of range.');
+	}
+
+	return value;
 
 }
